refactor(webapp-game): move off contribute joined projects move to actionMoves
